git-svn-id: svn://192.168.1.10/public/CoffeeMud@24322 0d6f1817-ed0e-0410-87c9-987e46238f29

+++ guides/database/coffeemuddb.sql
@@ -43,7 +43,7 @@ CREATE TABLE CMPOLL (
 	CMSUBJ char (255) NULL ,
 	CMDESC text NULL ,
 	CMOPTN text NULL ,
-	CMFLAG bigint NULL ,
+	CMFLAG int NULL ,
 	CMQUAL char (255) NULL ,
 	CMRESL text NULL,
 	CMEXPI bigint NULL
